Ver 0.4.u 	1. 사생활 관련 데이터가 길어서 UART 데이터가 2번으로 나눠서 전달되는 내용 수정 	 - 나눠 오는 경우 전체 패킷에 길이를 비교 후 처리하는 것으로 수정 	2. 사생활 관련 버그 수정 	 - 28개 데이터를 27개로 변경 	 - 데이터 배열이 int로 되어 있는 내용을 uint8_t 로 변경 	 - 문, 사용자에 대한 내용이 합쳐저 있는 내용을 분리 	3. 클립 동작 시 스트리밍 동작 버그 수정 및 기능 변경 	 - 클립 동작 후 스트리밍 연결 시 연결 안되는 내용 수정 	 - 클립 동작 후 스트리밍 연결 후 종료 시 클립 저장 종료 추가 	 - 저장 종료 후 정상 파일 전송 시작 	4. 택배 사라짐 기능 관련 버그 수정 	 - 택배 캡쳐를 보낸 후 오리지널 사진과 비교 후 테그를 변경하는 내용 	 - 택배 캡쳐 전에 오리지널 사진과 비굑하여 사라짐 동작 정상화 	5. 템퍼 동작 버그 수정 	 - 템퍼 동작 시 영상에 움직임이 없을 경우 클립과 같이 종료시켜버림 	 - 템퍼로 인한 동작 시 움직임과 관계없이 영상 저장 시작 	 - 종료 내용은 기존 클립과 동일(움직임 or 사람 or 얼굴 X, 3초)

+++ setting.h
@@ -31,8 +31,8 @@ typedef struct {
     int move_ex_s_y;
     int move_ex_e_x;
     int move_ex_e_y;
-    int door_grid[28];
-    int user_grid[28];
+    uint8_t door_grid[28];
+    uint8_t user_grid[28];
     SETTING_FLAGS SF;
 } Settings;
